split pubsub listener in two parts
- local cache refresher
- websocket publisher
