+++ src/common/pubsub/pub.sub.listener.controller.ts
@@ -1,9 +1,7 @@
 import { OriginLogger } from "@multiversx/sdk-nestjs";
 import { ElrondCachingService } from "@multiversx/sdk-nestjs";
-import { ShardTransaction } from "@elrondnetwork/transaction-processor";
 import { Controller } from "@nestjs/common";
 import { EventPattern } from "@nestjs/microservices";
-import { WebSocketPublisherService } from "src/common/websockets/web-socket-publisher-service";
 
 @Controller()
 export class PubSubListenerController {
@@ -11,7 +9,6 @@ export class PubSubListenerController {
 
   constructor(
     private readonly cachingService: ElrondCachingService,
-    private readonly webSocketPublisherService: WebSocketPublisherService,
   ) { }
 
   @EventPattern('deleteCacheKeys')
@@ -26,25 +23,5 @@ export class PubSubListenerController {
   async refreshCacheKey(info: { key: string, ttl: number }) {
     this.logger.log(`Refreshing local cache key ${info.key} with ttl ${info.ttl}`);
     await this.cachingService.refreshLocal(info.key, info.ttl);
-  }
-
-  @EventPattern('transactionsCompleted')
-  async transactionsCompleted(transactions: ShardTransaction[]) {
-    for (const transaction of transactions) {
-      await this.webSocketPublisherService.onTransactionCompleted(transaction);
-    }
-  }
-
-  @EventPattern('transactionsPendingResults')
-  async transactionsPendingResults(transactions: ShardTransaction[]) {
-    for (const transaction of transactions) {
-      await this.webSocketPublisherService.onTransactionPendingResults(transaction);
-    }
-  }
-
-  @EventPattern('onBatchUpdated')
-  async onBatchUpdated(payload: { address: string, batchId: string, txHashes: string[] }) {
-    this.logger.log(`Notifying batch updated for address ${payload.address}, batch id '${payload.batchId}', hashes ${payload.txHashes}`);
-    await this.webSocketPublisherService.onBatchUpdated(payload.address, payload.batchId, payload.txHashes);
   }
 }
