+++ src/common/pubsub/pub.sub.listener.module.ts
@@ -1,12 +1,10 @@
 import { Module } from '@nestjs/common';
 import { DynamicModuleUtils } from 'src/utils/dynamic.module.utils';
-import { WebSocketPublisherModule } from 'src/common/websockets/web-socket-publisher-module';
 import { PubSubListenerController } from './pub.sub.listener.controller';
 
 @Module({
   imports: [
     DynamicModuleUtils.getCachingModule(),
-    WebSocketPublisherModule,
   ],
   controllers: [
     PubSubListenerController,
